Create default instructions in MapboxNavigation and rename defaultMilestone boolean
